refactor: extract SkillsBadgesList and CertificatesPanel to a section
